Split the crazy-large Variables.cc file into two.
